[HOTSPOT-174] refactor: family 도메인 JPA/JDBC 역할 분리 및 리포지토리 구조 개선
