+++ src/main/java/hotspot/admin/family/infrastructure/FamilyApplyJpaRepository.java
@@ -1,0 +1,33 @@
+package hotspot.admin.family.infrastructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+
+import hotspot.admin.family.domain.ApplyType;
+import hotspot.admin.family.domain.FamilyApplyStatus;
+import hotspot.admin.family.infrastructure.entity.FamilyApplyEntity;
+
+public interface FamilyApplyJpaRepository extends JpaRepository<FamilyApplyEntity, Long> {
+    @Modifying(clearAutomatically = true, flushAutomatically = true)
+    @Query(
+            value = """
+                    UPDATE family_apply
+                    SET status = :newStatus,
+                        modified_time = now()
+                    WHERE family_apply_id = :familyApplyId
+                      AND apply_type = :applyType
+                      AND status = :currentStatus
+                    """,
+            nativeQuery = true
+    )
+    int updateStatusByIdAndTypeAndCurrentStatus(
+            @Param("familyApplyId") Long familyApplyId,
+            @Param("applyType") ApplyType applyType,
+            @Param("currentStatus") FamilyApplyStatus currentStatus,
+            @Param("newStatus") FamilyApplyStatus newStatus
+    );
+
+    boolean existsByFamilyApplyIdAndApplyType(Long familyApplyId, ApplyType applyType);
+}
