+++ src/main/java/hotspot/admin/family/infrastructure/FamilyJpaRepository.java
@@ -1,8 +1,22 @@
 package hotspot.admin.family.infrastructure;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+
+import hotspot.admin.family.domain.PriorityType;
 import hotspot.admin.family.infrastructure.entity.FamilyEntity;
 
 public interface FamilyJpaRepository extends JpaRepository<FamilyEntity, Long> {
+    boolean existsByFamilyIdAndIsDeletedFalse(Long familyId);
+
+    @Query("""
+            SELECT f.priorityType
+            FROM FamilyEntity f
+            WHERE f.familyId = :familyId
+              AND f.isDeleted = false
+            """)
+    Optional<PriorityType> findPriorityTypeByFamilyId(@Param("familyId") Long familyId);
 }
