+++ src/main/java/hotspot/admin/family/infrastructure/FamilySubJpaRepository.java
@@ -1,0 +1,8 @@
+package hotspot.admin.family.infrastructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import hotspot.admin.family.infrastructure.entity.FamilySubEntity;
+
+public interface FamilySubJpaRepository extends JpaRepository<FamilySubEntity, Long> {
+}
