+++ src/main/java/hotspot/admin/family/service/GetFamilyPolicyStatusServiceImpl.java
@@ -14,6 +14,7 @@ import hotspot.admin.family.controller.port.GetFamilyPolicyStatusService;
 import hotspot.admin.family.controller.response.FamilyPolicyMemberStatusItem;
 import hotspot.admin.family.service.dto.FamilyPolicyStatusRow;
 import hotspot.admin.family.service.port.FamilyRepository;
+import hotspot.admin.family.service.port.FamilySubRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -21,6 +22,7 @@ import lombok.RequiredArgsConstructor;
 public class GetFamilyPolicyStatusServiceImpl implements GetFamilyPolicyStatusService {
 
     private final FamilyRepository familyRepository;
+    private final FamilySubRepository familySubRepository;
     private final PhoneCryptoUtil phoneCryptoUtil;
 
     /** 구성원별 적용 정책(시간대/차단 서비스)과 차단 상태를 묶어 반환한다. */
@@ -31,7 +33,7 @@ public class GetFamilyPolicyStatusServiceImpl implements GetFamilyPolicyStatusSe
             throw new ApplicationException(FamilyErrorCode.FAMILY_NOT_FOUND);
         }
 
-        return familyRepository.findFamilyPolicyStatusRows(familyId).stream()
+        return familySubRepository.findFamilyPolicyStatusRows(familyId).stream()
                 .map(this::toMemberItem)
                 .toList();
     }
