+++ src/main/java/hotspot/admin/family/service/ProcessFamilyRequestServiceImpl.java
@@ -8,14 +8,14 @@ import hotspot.admin.common.exception.code.FamilyErrorCode;
 import hotspot.admin.family.controller.port.ProcessFamilyRequestService;
 import hotspot.admin.family.domain.ApplyType;
 import hotspot.admin.family.domain.FamilyApplyStatus;
-import hotspot.admin.family.service.port.FamilyRepository;
+import hotspot.admin.family.service.port.FamilyApplyRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class ProcessFamilyRequestServiceImpl implements ProcessFamilyRequestService {
 
-    private final FamilyRepository familyRepository;
+    private final FamilyApplyRepository familyApplyRepository;
 
     /** 가족 요청을 승인 상태로 전환한다. */
     @Override
@@ -37,7 +37,7 @@ public class ProcessFamilyRequestServiceImpl implements ProcessFamilyRequestServ
             Long requestId,
             FamilyApplyStatus nextStatus
     ) {
-        int updatedCount = familyRepository.updateFamilyRequestStatus(
+        int updatedCount = familyApplyRepository.updateFamilyRequestStatus(
                 requestId,
                 applyType,
                 FamilyApplyStatus.PENDING,
@@ -51,7 +51,7 @@ public class ProcessFamilyRequestServiceImpl implements ProcessFamilyRequestServ
 
     /** 업데이트 실패 시 요청 미존재/상태 불일치 원인을 구분해 예외를 던진다. */
     private void handleNotUpdated(Long requestId, ApplyType applyType) {
-        if (!familyRepository.existsFamilyRequest(requestId, applyType)) {
+        if (!familyApplyRepository.existsFamilyRequest(requestId, applyType)) {
             throw new ApplicationException(FamilyErrorCode.FAMILY_REQUEST_NOT_FOUND);
         }
         throw new ApplicationException(FamilyErrorCode.FAMILY_REQUEST_NOT_PENDING);
