+++ src/test/java/hotspot/admin/family/service/GetFamilyPolicyStatusServiceTest.java
@@ -21,6 +21,7 @@ import hotspot.admin.family.controller.response.FamilyPolicyMemberStatusItem;
 import hotspot.admin.family.domain.FamilyRole;
 import hotspot.admin.family.service.dto.FamilyPolicyStatusRow;
 import hotspot.admin.family.service.port.FamilyRepository;
+import hotspot.admin.family.service.port.FamilySubRepository;
 
 @ExtendWith(MockitoExtension.class)
 class GetFamilyPolicyStatusServiceTest {
@@ -29,20 +30,23 @@ class GetFamilyPolicyStatusServiceTest {
     private FamilyRepository familyRepository;
 
     @Mock
+    private FamilySubRepository familySubRepository;
+
+    @Mock
     private PhoneCryptoUtil phoneCryptoUtil;
 
     private GetFamilyPolicyStatusServiceImpl service;
 
     @BeforeEach
     void setUp() {
-        service = new GetFamilyPolicyStatusServiceImpl(familyRepository, phoneCryptoUtil);
+        service = new GetFamilyPolicyStatusServiceImpl(familyRepository, familySubRepository, phoneCryptoUtil);
     }
 
     @Test
     @DisplayName("가족 정책 적용 현황 조회 성공")
     void getFamilyPolicyStatusSuccess() throws Exception {
         when(familyRepository.existsFamilyById(1L)).thenReturn(true);
-        when(familyRepository.findFamilyPolicyStatusRows(1L))
+        when(familySubRepository.findFamilyPolicyStatusRows(1L))
                 .thenReturn(List.of(
                         FamilyPolicyStatusRow.builder()
                                 .subId(10L)
@@ -90,7 +94,7 @@ class GetFamilyPolicyStatusServiceTest {
     @DisplayName("전화번호 복호화 실패 시 예외")
     void decryptFailThenException() throws Exception {
         when(familyRepository.existsFamilyById(1L)).thenReturn(true);
-        when(familyRepository.findFamilyPolicyStatusRows(1L))
+        when(familySubRepository.findFamilyPolicyStatusRows(1L))
                 .thenReturn(List.of(FamilyPolicyStatusRow.builder()
                         .subId(1L)
                         .memberName("대표")
