+++ src/test/java/hotspot/admin/family/service/ProcessFamilyRequestServiceTest.java
@@ -15,25 +15,25 @@ import hotspot.admin.common.exception.ApplicationException;
 import hotspot.admin.common.exception.code.FamilyErrorCode;
 import hotspot.admin.family.domain.ApplyType;
 import hotspot.admin.family.domain.FamilyApplyStatus;
-import hotspot.admin.family.service.port.FamilyRepository;
+import hotspot.admin.family.service.port.FamilyApplyRepository;
 
 @ExtendWith(MockitoExtension.class)
 class ProcessFamilyRequestServiceTest {
 
     @Mock
-    private FamilyRepository familyRepository;
+    private FamilyApplyRepository familyApplyRepository;
 
     private ProcessFamilyRequestServiceImpl service;
 
     @BeforeEach
     void setUp() {
-        service = new ProcessFamilyRequestServiceImpl(familyRepository);
+        service = new ProcessFamilyRequestServiceImpl(familyApplyRepository);
     }
 
     @Test
     @DisplayName("대기중 신청 요청 승인 성공")
     void approveSuccess() {
-        when(familyRepository.updateFamilyRequestStatus(
+        when(familyApplyRepository.updateFamilyRequestStatus(
                 7L,
                 ApplyType.ADD,
                 FamilyApplyStatus.PENDING,
@@ -42,7 +42,7 @@ class ProcessFamilyRequestServiceTest {
 
         service.approve(ApplyType.ADD, 7L);
 
-        verify(familyRepository).updateFamilyRequestStatus(
+        verify(familyApplyRepository).updateFamilyRequestStatus(
                 7L,
                 ApplyType.ADD,
                 FamilyApplyStatus.PENDING,
@@ -53,13 +53,13 @@ class ProcessFamilyRequestServiceTest {
     @Test
     @DisplayName("요청이 존재하지 않으면 예외")
     void requestNotFoundThenException() {
-        when(familyRepository.updateFamilyRequestStatus(
+        when(familyApplyRepository.updateFamilyRequestStatus(
                 9L,
                 ApplyType.REMOVE,
                 FamilyApplyStatus.PENDING,
                 FamilyApplyStatus.REJECTED
         )).thenReturn(0);
-        when(familyRepository.existsFamilyRequest(9L, ApplyType.REMOVE))
+        when(familyApplyRepository.existsFamilyRequest(9L, ApplyType.REMOVE))
                 .thenReturn(false);
 
         assertThatThrownBy(() -> service.reject(ApplyType.REMOVE, 9L))
@@ -70,13 +70,13 @@ class ProcessFamilyRequestServiceTest {
     @Test
     @DisplayName("이미 처리된 요청이면 예외")
     void requestNotPendingThenException() {
-        when(familyRepository.updateFamilyRequestStatus(
+        when(familyApplyRepository.updateFamilyRequestStatus(
                 9L,
                 ApplyType.REMOVE,
                 FamilyApplyStatus.PENDING,
                 FamilyApplyStatus.REJECTED
         )).thenReturn(0);
-        when(familyRepository.existsFamilyRequest(9L, ApplyType.REMOVE))
+        when(familyApplyRepository.existsFamilyRequest(9L, ApplyType.REMOVE))
                 .thenReturn(true);
 
         assertThatThrownBy(() -> service.reject(ApplyType.REMOVE, 9L))
